Need to modify this file
The intent is to add recesses for the LED and photodiode to fit. Use a tolerance of 1mm in both axes parallel to the plane. e.g. for a cross-section of 3mmx5mm, use 4mmx6mm. This is to account for the 3D rprinter's imprecision. Please add the dimension as the desired dimension plus the tolerance, e.g. '3mm + 1mm' because it helps when fine tuning the parameters later.